(feat) add Title component with screenreader feature
