+++ src/components/Title/TitleStyles.js
@@ -1,0 +1,3 @@
+import { ScreenReaderTitleStyles } from "gertrudenyenyeshi/styles";
+
+export const ScreenReaderTitle = ScreenReaderTitleStyles;
